Ensure values are of float

+++ wosis/analysis/stats.py
@@ -24,7 +24,7 @@ def calc_average_citations(in_df, sort=False):
         'DataFrame has to have `citation` column. Use `get_num_citations()` first'
     out_df = in_df.copy()
     max_year = out_df.year.max()
-    out_df.loc[:, 'Avg. Citations'] = (out_df.citations / ((max_year - out_df.year) + 1)).round(2)
+    out_df.loc[:, 'Avg. Citations'] = (out_df.citations / ((max_year - out_df.year) + 1)).astype(float).round(2)
 
     if sort:
         out_df = out_df.sort_values('Avg. Citations', ascending=False)
